Obey Android 12 requirement
android:exported
Whether the activity can be launched by components of other applications:

If "true", the activity is accessible to any app, and is launchable by its exact class name.
If "false", the activity can be launched only by components of the same application, applications with the same user ID, or privileged system components. This is the default value when there are no intent filters.
If an activity in your app includes intent filters, set this element to "true" to let other apps start it. For example, if the activity is the main activity of the app and includes the category android.intent.category.LAUNCHER.

If this element is set to "false" and an app tries to start the activity, the system throws an ActivityNotFoundException.

This attribute isn't the only way to limit an activity's exposure to other applications. Permissions are also used to limit the external entities that can invoke the activity. See the permission attribute.